maquette-py/pyproject.toml
[tool.poetry]
name = "maquette"
version = "0.0.7"
description = "A command line interface for Maquette Data Science and Machine Learning Platform."
authors = ["Johannes Forster <johannesforster@kpmg.com>", "Michael Wellner <michaelwellner@kpmg.com>"]
license = "Apache-2.0"
readme = "Readme.md"

packages = [
            {include = "maquette"},
            {include = "maquette_lib"}]

[tool.poetry.dependencies]
python = "~3.8"
click = "^7.1.2"
requests = "^2.24.0"
pandas = "^1.1.0"
pandavro = "^1.5.2"
seaborn = "^0.11.0"
pyyaml = "^5.4.1"
gitpython = "^3.1.12"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"